BC-9277 - remove micro-orm folder in repo folder from account module
