+++ apps/server/src/modules/account/repo/account.repo.integration.spec.ts
@@ -5,9 +5,9 @@ import { userFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { MongoMemoryDatabaseModule } from '@testing/database';
-import { Account, ACCOUNT_REPO } from '../../domain';
-import { AccountEntity } from '../../domain/entity/account.entity';
-import { accountDoFactory, accountFactory } from '../../testing';
+import { Account, ACCOUNT_REPO } from '../domain';
+import { AccountEntity } from '../domain/entity/account.entity';
+import { accountDoFactory, accountFactory } from '../testing';
 import { AccountMikroOrmRepo } from './account.repo';
 import { AccountEntityToDoMapper } from './mapper';
 import { AccountDoToEntityMapper } from './mapper/account-do-to-entity.mapper';
