+++ apps/server/src/modules/account/repo/account.repo.ts
@@ -4,9 +4,9 @@ import { Injectable } from '@nestjs/common';
 import { SortOrder } from '@shared/domain/interface';
 import { Counted, EntityId } from '@shared/domain/types';
 import { BaseDomainObjectRepo } from '@shared/repo/base-domain-object.repo';
-import { AccountRepo } from '../../domain';
-import { Account } from '../../domain/account';
-import { AccountEntity } from '../../domain/entity/account.entity';
+import { AccountRepo } from '../domain';
+import { Account } from '../domain/account';
+import { AccountEntity } from '../domain/entity/account.entity';
 import { AccountEntityToDoMapper } from './mapper';
 import { AccountDoToEntityMapper } from './mapper/account-do-to-entity.mapper';
 import { AccountScope } from './scope/account-scope';
