+++ apps/server/src/modules/account/repo/index.ts
@@ -1,0 +1,2 @@
+export { AccountMikroOrmRepo } from './account.repo';
+export * from './mapper';
