+++ apps/server/src/modules/account/repo/mapper/account-do-to-entity.mapper.spec.ts
@@ -1,5 +1,5 @@
 import { ObjectId } from '@mikro-orm/mongodb';
-import { accountDoFactory } from '../../../testing';
+import { accountDoFactory } from '../../testing';
 import { AccountDoToEntityMapper } from './account-do-to-entity.mapper';
 
 describe('AccountEntityToDoMapper', () => {
