+++ apps/server/src/modules/account/repo/mapper/account-idm-to-do.mapper.db.spec.ts
@@ -1,37 +1,31 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { IdmAccount } from '@shared/domain/interface';
-import { Account } from '../../../domain';
+import { Account } from '../../domain';
 import { AccountIdmToDoMapper } from './account-idm-to-do.mapper.abstract';
-import { AccountIdmToDoMapperIdm } from './account-idm-to-do.mapper.idm';
+import { AccountIdmToDoMapperDb } from './account-idm-to-do.mapper.db';
 
-describe('AccountIdmToDoMapperIdm', () => {
+describe('AccountIdmToDoMapperDb', () => {
 	let module: TestingModule;
 	let mapper: AccountIdmToDoMapper;
-
-	const now: Date = new Date(2022, 1, 22);
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
 			providers: [
 				{
 					provide: AccountIdmToDoMapper,
-					useClass: AccountIdmToDoMapperIdm,
+					useClass: AccountIdmToDoMapperDb,
 				},
 			],
 		}).compile();
 
 		mapper = module.get(AccountIdmToDoMapper);
-
-		jest.useFakeTimers();
-		jest.setSystemTime(now);
 	});
 
 	afterAll(async () => {
 		await module.close();
 	});
-
 	describe('mapToDo', () => {
-		describe('when mapping from entity to do', () => {
+		describe('when mapping from entity to dto', () => {
 			const setup = () => {
 				const testIdmEntity: IdmAccount = {
 					id: 'id',
@@ -54,8 +48,8 @@ describe('AccountIdmToDoMapperIdm', () => {
 
 				expect(ret).toEqual(
 					expect.objectContaining<Partial<Account>>({
-						id: testIdmEntity.id,
-						idmReferenceId: undefined,
+						id: testIdmEntity.attDbcAccountId,
+						idmReferenceId: testIdmEntity.id,
 						userId: testIdmEntity.attDbcUserId,
 						systemId: testIdmEntity.attDbcSystemId,
 						createdAt: testIdmEntity.createdDate,
@@ -65,39 +59,48 @@ describe('AccountIdmToDoMapperIdm', () => {
 				);
 			});
 		});
+
 		describe('when date is undefined', () => {
 			const setup = () => {
 				const testIdmEntity: IdmAccount = {
 					id: 'id',
 				};
-				return testIdmEntity;
+
+				const dateMock = new Date();
+				jest.useFakeTimers();
+				jest.setSystemTime(dateMock);
+
+				return { testIdmEntity, dateMock };
 			};
 
 			it('should use actual date', () => {
-				const testIdmEntity = setup();
+				const { testIdmEntity, dateMock } = setup();
 
 				const ret = mapper.mapToDo(testIdmEntity);
 
-				expect(ret.createdAt).toEqual(now);
-				expect(ret.updatedAt).toEqual(now);
+				expect(ret.createdAt).toEqual(dateMock);
+				expect(ret.updatedAt).toEqual(dateMock);
+
+				jest.useRealTimers();
 			});
 		});
+	});
 
-		describe('when a fields value is missing', () => {
-			const setup = () => {
-				const testIdmEntity: IdmAccount = {
-					id: 'id',
-				};
-				return testIdmEntity;
+	describe('when a fields value is missing', () => {
+		const setup = () => {
+			const testIdmEntity: IdmAccount = {
+				id: 'id',
 			};
+			return testIdmEntity;
+		};
 
-			it('should fill with empty string', () => {
-				const testIdmEntity = setup();
+		it('should fill with empty string', () => {
+			const testIdmEntity = setup();
 
-				const ret = mapper.mapToDo(testIdmEntity);
+			const ret = mapper.mapToDo(testIdmEntity);
 
-				expect(ret.username).toBe('');
-			});
+			expect(ret.id).toBe('');
+			expect(ret.username).toBe('');
 		});
 	});
 });
